+++ examples/todoapp/integration_test/main/base/tags.dart
@@ -1,0 +1,3 @@
+const String regressionTest = 'regression';
+const String negativeTest = 'negative';
+const String positiveTest = 'positive';