+++ packages/rx_bloc_cli/mason_templates/bricks/lib_permissions/__brick__/lib/lib_permissions/models/route_permissions.dart
@@ -3,8 +3,8 @@
 class RoutePermissions { {{#enable_feature_qr_scanner}}
   static const qrCode = 'QrCodeRoute';{{/enable_feature_qr_scanner}} {{#enable_mfa}}
   static const mfa = 'MfaRoute';{{/enable_mfa}}
-  static const dashboard = 'DashboardRoute';
-  static const profile = 'ProfileRoute';
+  static const dashboard = 'DashboardRoute';{{#enable_profile}}
+  static const profile = 'ProfileRoute';{{/enable_profile}}
   static const splash = 'SplashRoute';{{#enable_feature_counter}}
   static const counter = 'CounterRoute';{{/enable_feature_counter}}{{#enable_feature_widget_toolkit}}
   static const widgetToolkit = 'WidgetToolkitRoute';{{/enable_feature_widget_toolkit}}