+++ packages/rx_bloc_cli/mason_templates/bricks/lib_router/__brick__/lib/lib_router/routes/routes.dart
@@ -65,6 +65,7 @@ class SplashRoute extends GoRouteData implements RouteDataModel {
         ),
       ],
     ),{{/enable_mfa}}
+    {{#enable_profile}}
     TypedStatefulShellBranch<ProfileBranchData>(
       routes: <TypedRoute<RouteData>>[
         TypedGoRoute<ProfileRoute>(
@@ -82,7 +83,7 @@ class SplashRoute extends GoRouteData implements RouteDataModel {
           ],
         ),
       ],
-    ),
+    ),{{/enable_profile}}
   ],
 )
 @immutable
@@ -131,3 +132,24 @@ class DashboardRoute extends GoRouteData implements RouteDataModel {
   @override
   String get routeLocation => location;
 }
+
+{{^enable_profile}}
+@immutable
+@TypedGoRoute<NotificationsRoute>(path: RoutesPath.notifications)
+class NotificationsRoute extends GoRouteData implements RouteDataModel {
+  const NotificationsRoute();
+
+  @override
+  Page<Function> buildPage(BuildContext context, GoRouterState state) =>
+      MaterialPage(
+        key: state.pageKey,
+        child: const NotificationsPageWithDependencies(),
+      );
+
+  @override
+  String get permissionName => RouteModel.notifications.permissionName;
+
+  @override
+  String get routeLocation => location;
+}
+{{/enable_profile}}