+++ packages/rx_bloc_cli/test/models/generator_arguments_test.dart
@@ -42,6 +42,7 @@ void main() {
       cicdGithubEnabled: true,
       cicdCodemagicEnabled: true,
       qrScannerEnabled: true,
+      profileEnabled: true,
     );
     sut = GeneratorArguments(
       outputDirectory: outputDirectory,
@@ -93,6 +94,7 @@ void main() {
           equals(featureConfiguration.cicdCodemagicEnabled));
       expect(
           sut.qrScannerEnabled, equals(featureConfiguration.qrScannerEnabled));
+      expect(sut.profileEnabled, equals(featureConfiguration.profileEnabled));
     });
   });
 }