+++ packages/rx_bloc_cli/test/stub.dart
@@ -61,6 +61,19 @@ final class Stub {
     ..[CreateCommandArguments.socialLogins.name] = false
     ..[CreateCommandArguments.otp.name] = false;
 
+  static Map<String, Object> get changeLanguageEnabled =>
+      Map.from(Stub.defaultValues)
+        ..[CreateCommandArguments.changeLanguage.name] = true
+        ..[CreateCommandArguments.mfa.name] = false
+        ..[CreateCommandArguments.login.name] = false
+        ..[CreateCommandArguments.socialLogins.name] = false
+        ..[CreateCommandArguments.otp.name] = false
+        ..[CreateCommandArguments.profile.name] = false;
+
+  static Map<String, Object> get loginEnabled => Map.from(Stub.defaultValues)
+    ..[CreateCommandArguments.login.name] = true
+    ..[CreateCommandArguments.profile.name] = false;
+
   static final generatorArgumentsAllEnabled = GeneratorArguments(
     outputDirectory: Directory('some/output_directory'),
     projectConfiguration: ProjectConfiguration(
@@ -88,6 +101,7 @@ final class Stub {
       cicdGithubEnabled: true,
       cicdCodemagicEnabled: true,
       qrScannerEnabled: true,
+      profileEnabled: true,
     ),
   );
 
@@ -123,6 +137,7 @@ final class Stub {
           CreateCommandArguments.cicd.defaultValue() == CICDType.github,
       cicdCodemagicEnabled:
           CreateCommandArguments.cicd.defaultValue() == CICDType.codemagic,
+      profileEnabled: CreateCommandArguments.profile.defaultValue(),
     ),
   );
 }